backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
numpy==2.0.1
pandas==2.2.2
joblib==1.4.2
tensorflow==2.18.0
yfinance==0.2.43
requests==2.32.3
python-dotenv==1.0.1
scikit-learn==1.5.1
h5py==3.11.0
